requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.37.0
urllib3==2.5.0
alembic==1.16.5
prometheus-fastapi-instrumentator==7.1.0
loguru==0.7.3
python-dotenv==1.1.1
aiohttp==3.12.15
httpx==0.28.1
python-jose[cryptography]==3.5.0
asyncpg==0.30.0
greenlet==3.2.4
psycopg[binary]==3.2.10 
pymongo==4.*
pydantic-settings==2.11.0
psycopg2-binary
email-validator
privy
requests
cryptography
PyJWT 
supabase
firebase-admin
pytest